+++ src/main/java/uk/gov/hmcts/opal/service/hmrc/clients/HmrcAuthClient.java
@@ -2,8 +2,7 @@ package uk.gov.hmcts.opal.service.hmrc.clients;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
-import uk.gov.hmcts.opal.service.hmrc.HmrcAuthCreds;
-import uk.gov.hmcts.opal.service.hmrc.response.HmrcAuthToken;
+import uk.gov.hmcts.opal.service.hmrc.clients.response.HmrcAuthToken;
 
 @FeignClient(name = "hmrcAuthClient", url = "${hmrc.auth.url}")
 public interface HmrcAuthClient {
